Remote source origin is parsed and extracted from SDP when the source is discovered or updated. This way we avoid parsing the SDP multiple times.
