+++ daemon/session_manager.hpp
@@ -32,23 +32,6 @@
 #include "igmp.hpp"
 #include "sap.hpp"
 
-struct SDPOrigin {
-  std::string username;
-  std::string session_id;
-  uint64_t session_version{0};
-  std::string network_type;
-  std::string address_type;
-  std::string unicast_address;
-
-  bool operator==(const SDPOrigin& rhs) const {
-    // session_version is not part of comparison, see RFC 4566
-    return username == rhs.username && session_id == rhs.session_id &&
-           network_type == rhs.network_type &&
-           address_type == rhs.address_type &&
-           unicast_address == rhs.unicast_address;
-  }
-};
-
 struct StreamSource {
   uint8_t id{0};
   bool enabled{false};
@@ -211,7 +194,6 @@ class SessionManager {
                            const std::list<RemoteSource> sources_list) const;
 
   bool parse_sdp(const std::string sdp, StreamInfo& info) const;
-  bool parse_sdp_origin(const std::string sdp, SDPOrigin& origin) const;
   bool worker();
   // singleton, use create() to build
   SessionManager(std::shared_ptr<DriverManager> driver,
